+++ editoast/src/views/authz.rs
@@ -2,7 +2,6 @@ use std::collections::HashMap;
 use std::collections::HashSet;
 use std::sync::Arc;
 
-use crate::authorizers::SystemAuthorizer;
 use crate::error::Result;
 use crate::views::authz::resources::Resource;
 use crate::views::authz::resources::StandardGrant;
@@ -13,6 +12,7 @@ use ::authz::InfraPrivilege;
 use ::authz::Role;
 use authz::RollingStockGrant;
 use authz::RollingStockPrivilege;
+use authz::SystemAuthorizer;
 use authz::v2;
 use authz::v2::Actor;
 use authz::v2::Authorizer;
