+++ editoast/src/views/infra/mod.rs
@@ -42,8 +42,6 @@ use super::pagination::PaginationStats;
 use crate::AppState;
 use crate::Arc;
 use crate::authentication;
-use crate::authorizers::SystemAuthorizer;
-use crate::authorizers::UserAuthorizer;
 use crate::error::Result;
 use crate::generated_data::InfraGeneratedData as _;
 use crate::generated_data::operational_point::OperationalPointLayer;
@@ -56,6 +54,8 @@ use crate::views::pagination::PaginationQueryParams;
 use crate::views::params;
 use crate::views::path::operational_point_cache::OperationalPointCache;
 use authz::Role;
+use authz::SystemAuthorizer;
+use authz::UserAuthorizer;
 use models::Infra;
 use models::SwitchTypeModel;
 use schemas::infra::OperationalPoint;
