+++ editoast/src/views/rolling_stock.rs
@@ -50,9 +50,9 @@ use utoipa::ToSchema;
 
 use crate::AppState;
 use crate::authentication;
-use crate::authorizers::SystemAuthorizer;
 use crate::error::InternalError;
 use crate::error::Result;
+use authz::SystemAuthorizer;
 
 #[derive(Debug, Serialize, ToSchema)]
 pub struct RollingStockWithLiveries {
