+++ editoast/src/views/rolling_stock/light.rs
@@ -31,11 +31,11 @@ use super::RollingStockIdParam;
 use super::RollingStockKey;
 use super::RollingStockNameParam;
 use crate::AppState;
-use crate::authorizers::SystemAuthorizer;
 use crate::error::Result;
 use crate::views::pagination::PaginatedList;
 use crate::views::pagination::PaginationQueryParams;
 use crate::views::pagination::PaginationStats;
+use authz::SystemAuthorizer;
 
 #[cfg(test)]
 use serde::Deserialize;
